split up tests suite
